# openapi/openapi.yaml
openapi: 3.1.1
info:
  title: Yuva API
  version: 0.0.3
  summary: Customer messaging for many products in one inbox.
  license:
    name: MIT
    identifier: MIT
  description: |
    Yuva answers every user of every product you run: e-mail, live chat and in-app
    conversations in one inbox.

    ## Surfaces

    - `/v1` is for members (panel session) and workspace API keys.
    - `/client/v1` is for the web widget and the mobile SDKs.
    - `/healthz` and `/readyz` are for load balancers and orchestrators.

    ## Errors

    Errors are `application/problem+json` documents (RFC 9457) with a stable `code`.

    ## Authentication

    - Members sign in with a one-time code sent by e-mail or with a passkey. Both set the
      `yuva_session` cookie (HttpOnly, SameSite=Lax).
    - Host backends, scripts and bots send a workspace API key as `Authorization: Bearer <key>`.
      A key acts on its own workspace only, and only with its scopes (see `api-keys`).
    - Contacts (widget, SDKs) send a contact session token as `Authorization: Bearer <token>` on
      `/client/v1`; see the `client` tag.

    ## Idempotency

    Every authenticated `POST` on `/v1` and `/client/v1` accepts an `Idempotency-Key` header (1 to
    255 printable ASCII characters). The key is remembered for 24 hours per caller (API key,
    member or contact) with the method, path, a SHA-256 of the body and the response. The same key
...
    rather than a workspace (`/v1/me/...`, `/v1/auth/...`), ignore the header.

...

...
servers:
...
security:
  - session: []
  - apiKey: []
...
paths:
...
  /v1/auth/code:
    post:
      operationId: requestSignInCode
      summary: Request a sign-in code
      description: |
        Sends a one-time code to the address when it belongs to a member or has a pending invite.
        The answer is the same either way and takes the same time (the mail is sent after it), so
        it does not reveal whether an address is known. Codes expire after 10 minutes; only the
        most recent code of an address is accepted.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInCodeRequest"
      responses:
        "202":
          description: Accepted. A code is on its way if the address is known.
        "400":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
  /v1/auth/code/verify:
    post:
      operationId: verifySignInCode
      summary: Sign in with a code
      description: |
        Exchanges a valid code for a member session. Each code accepts five attempts. After 20
        wrong codes for an address within 24 hours, codes for it are refused
        (`429 sign_in_paused`) until older failures leave that window, and the member is told by
        e-mail; passkeys still work. Pending invites for the address are accepted on success.
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInCodeVerify"
      responses:
        "200":
          $ref: "#/components/responses/SignedIn"
        "400":
          $ref: "#/components/responses/Problem"
        "429":
          $ref: "#/components/responses/Problem"
...
  /v1/auth/passkey:
    post:
      operationId: finishPasskeySignIn
      summary: Sign in with a passkey
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeyCeremonyResult"
      responses:
        "200":
          $ref: "#/components/responses/SignedIn"
        "400":
          $ref: "#/components/responses/Problem"
  /v1/auth/sign-out:
    post:
      operationId: signOut
      summary: Sign out
      description: Ends the current member session and clears the cookie.
      tags: [auth]
      security:
        - session: []
      responses:
        "204":
          description: Signed out.
          headers:
            Set-Cookie:
              description: Clears `yuva_session`.
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Problem"
  /v1/me:
    get:
      operationId: getMe
      summary: The signed-in person
      description: The person behind the session and the workspaces they belong to.
      tags: [me]
      security:
        - session: []
      responses:
        "200":
          description: The person and their memberships.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateMe
      summary: Update the signed-in person
      tags: [me]
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeUpdate"
      responses:
        "200":
          description: The updated person and their memberships.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteMe
      summary: Delete the signed-in person's account
      description: |
        Removes the person with their memberships, sessions, passkeys, push subscriptions and
        sign-in codes, and clears the cookie. Messages they wrote stay in their conversations
        without an author (`author.member_id` absent). Refused with `409 last_owner` while the
        person is the only owner of a workspace: make someone else an owner or delete that
        workspace first. Needs a member session; the body repeats the person's e-mail address.
      tags: [me]
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountDeletion"
      responses:
        "204":
          description: Deleted.
          headers:
            Set-Cookie:
              description: Clears `yuva_session`.
              schema:
                type: string
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/me/passkeys:
    get:
      operationId: listPasskeys
      summary: List passkeys
      tags: [me]
      security:
        - session: []
      responses:
        "200":
          description: The person's passkeys.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyList"
        "401":
          $ref: "#/components/responses/Problem"
    post:
      operationId: finishPasskeyRegistration
      summary: Register a passkey
      description: Completes a registration started with `POST /v1/me/passkeys/options`.
      tags: [me]
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasskeyRegistration"
      responses:
        "201":
          description: The registered passkey.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passkey"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
  /v1/me/passkeys/options:
    post:
      operationId: beginPasskeyRegistration
      summary: Start a passkey registration
      description: |
        Returns the options for `navigator.credentials.create()` and a ceremony id to send back
        with the result. The ceremony expires after 5 minutes.
      tags: [me]
      security:
        - session: []
      responses:
        "200":
          description: The ceremony options.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasskeyCeremony"
        "401":
          $ref: "#/components/responses/Problem"
  /v1/me/passkeys/{passkeyId}:
    delete:
      operationId: deletePasskey
      summary: Remove a passkey
      tags: [me]
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/PasskeyId"
      responses:
        "204":
          description: Removed.
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/push/vapid-public-key:
    get:
      operationId: getVapidPublicKey
      summary: The server's Web Push key
      description: |
        The VAPID public key to pass as `applicationServerKey` to `PushManager.subscribe()`: an
        uncompressed P-256 point, base64url without padding. 404 `push_disabled` when the server
        has no VAPID keys configured; hide push settings then.
      tags: [notifications]
      security:
        - session: []
      responses:
        "200":
          description: The key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VapidPublicKey"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/me/push-subscriptions:
    get:
      operationId: listPushSubscriptions
      summary: List push subscriptions
      description: The signed-in person's push subscriptions, newest first.
      tags: [notifications]
      security:
        - session: []
      responses:
        "200":
          description: The subscriptions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscriptionList"
        "401":
          $ref: "#/components/responses/Problem"
    post:
      operationId: createPushSubscription
      summary: Register a push subscription
      description: |
        Saves the browser's `PushSubscription` (`subscription.toJSON()`) for the current session.
        Registering an endpoint that is already saved updates it and moves it to this session and
        person, so call it on every panel start while notification permission is granted. The
        endpoint must be an `https` URL on a public address. At most 20 subscriptions per person.
        404 `push_disabled` when the server has no VAPID keys.
      tags: [notifications]
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionCreate"
      responses:
        "201":
          description: The saved subscription.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushSubscription"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/me/push-subscriptions/{pushSubscriptionId}:
    delete:
      operationId: deletePushSubscription
      summary: Remove a push subscription
      description: Call it before `PushSubscription.unsubscribe()` when the member turns push off.
      tags: [notifications]
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/PushSubscriptionId"
      responses:
        "204":
          description: Removed.
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/me/push-subscriptions/{pushSubscriptionId}/test:
    post:
      operationId: testPushSubscription
      summary: Send a test notification
      description: |
        Queues a test notification to the subscription and returns its payload. The outcome shows
        up as `last_success_at` or `last_error` on the subscription.
      tags: [notifications]
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/PushSubscriptionId"
      responses:
        "202":
          description: Queued; the push carries this payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushNotification"
        "401":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/me/notifications:
    get:
      operationId: getNotificationSettings
      summary: Your notification settings
      description: The calling member's notification settings in the workspace. Member sessions only.
      tags: [notifications]
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
      responses:
        "200":
          description: The settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateNotificationSettings
      summary: Change your notification settings
      description: |
        Each event sent replaces that event's channels; events not sent keep theirs. Member
        sessions only.
      tags: [notifications]
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSettingsUpdate"
      responses:
        "200":
          description: The settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
  /v1/me/notifications/inboxes/{inboxId}:
    put:
      operationId: setInboxNotifications
      summary: Override your notifications for one inbox
      description: |
        Replaces the member's override for the inbox: the events sent here win over the
        workspace-wide settings in that inbox; the others follow them. Member sessions only.
      tags: [notifications]
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
        - $ref: "#/components/parameters/InboxId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InboxNotificationsUpdate"
      responses:
        "200":
          description: The override.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxNotifications"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteInboxNotifications
      summary: Remove your override for one inbox
      description: The inbox follows the workspace-wide settings again. Member sessions only.
      tags: [notifications]
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
        - $ref: "#/components/parameters/InboxId"
      responses:
        "204":
          description: Removed.
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
  /v1/workspace:
    get:
      operationId: getWorkspace
      summary: The current workspace
      description: "Scope: `inboxes:read`."
      tags: [workspace]
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
      responses:
        "200":
          description: The workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateWorkspace
      summary: Update the workspace
      description: |
        `retention_days`: owners, or an API key with `workspace:manage`. `bots_may_send`: owners
        and admins with a member session (`403 member_session_required` for a key).

        Scope: `workspace:manage`.
      tags: [workspace]
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdate"
      responses:
        "200":
          description: The updated workspace.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteWorkspace
      summary: Delete the workspace
      description: |
        Owners only, with a member session; the body repeats the workspace's name exactly. The
        workspace stops answering at once: its members, API keys, widget and app sessions and
        e-mail addresses are refused from this request on, its webhook endpoints are removed and
        nothing more is sent to them. A background job then deletes everything else in batches:
        conversations, messages, attachments and raw e-mails with their stored files, contacts,
        inboxes, channels, members, invites and settings. People keep their account; members left
        with no other workspace lose their push subscriptions.
      tags: [workspace]
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceDeletion"
      responses:
        "202":
          description: The workspace is closed and its data is being deleted.
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
...
  /v1/members/{memberId}:
    get:
      operationId: getMember
      summary: Get a member
      description: "Scope: `inboxes:read`."
      tags: [workspace]
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
        - $ref: "#/components/parameters/MemberId"
      responses:
        "200":
          description: The member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updateMember
      summary: Change a member's role
      description: |
        Owners and admins only, with a member session. Only an owner grants or takes away the
        `owner` role. The last owner cannot be demoted.
      tags: [workspace]
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
        - $ref: "#/components/parameters/MemberId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: The updated member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
    delete:
      operationId: removeMember
      summary: Remove a member
      description: |
        Owners and admins only, with a member session; only an owner removes an owner. The last
        owner cannot be removed.
      tags: [workspace]
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
        - $ref: "#/components/parameters/MemberId"
      responses:
        "204":
          description: Removed.
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
...
  /v1/invites/{inviteId}:
    delete:
      operationId: deleteInvite
      summary: Withdraw an invite
      tags: [workspace]
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceHeader"
        - $ref: "#/components/parameters/InviteId"
      responses:
        "204":
          description: Withdrawn.
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
...
  /client/v1/conversations/{conversationId}:
    get:
      operationId: getClientConversation
      summary: Get one of the contact's conversations
      description: Another contact's conversation, or one of another inbox, answers `404`.
      tags: [client]
      security:
        - contactSession: []
      parameters:
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "200":
          description: The conversation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientConversation"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: yuva_session
      description: |
        Member session set by a successful sign-in. Requests other than `GET` and `HEAD` with this
        cookie must send an `Origin` of the server's public origin (or a WebAuthn origin), or
        `Sec-Fetch-Site: same-origin` without `Origin` (`403 origin_not_allowed`), and a body as
        `application/json` or `multipart/form-data` (`415 unsupported_media_type`).
    apiKey:
      type: http
      scheme: bearer
      description: |
        A workspace API key, `yuva_<prefix>_<secret>`. Each operation a key may call needs one
        scope (`ApiKeyScope`); a call without it is refused with `403 insufficient_scope` naming the
        missing scope in `scope`. Operations that list no scope need a member session
        (`403 member_session_required`). An expired key is refused with `401 api_key_expired`.
    contactSession:
      type: http
      scheme: bearer
      description: A contact session token from `POST /client/v1/session`, `yuva_cs_<secret>`.
...